+++ src/app/_components/HeroSection/index.tsx
@@ -9,6 +9,7 @@ import { Suspense, useEffect, useRef, useState } from "react";
 import useReducedMotion from "@/app/_hooks/useReducedMotion";
 import useViewport from "@/app/_hooks/useViewport";
 import { fadeUp } from "@/lib/animation/variants";
+import { imageSize } from "@/lib/image-sizes";
 import { cn } from "@/lib/utils";
 
 const Hero3DCanvas = dynamic(() => import("./Hero3DCanvas"), {
