+++ src/app/_components/TeamMember/index.tsx
@@ -4,7 +4,9 @@ import type { User } from "@prisma/client";
 import Image from "next/image";
 import Link from "next/link";
 import defaultProfilePicture from "public/assets/DefaultProfilePicture.png";
+import { useState } from "react";
 import { formatDateOnly } from "@/app/_lib/utils";
+import { imageSize } from "@/lib/image-sizes";
 import { cn } from "@/lib/utils";
 
 import Linkedin from "../svgs/Linkedin";
@@ -29,23 +31,20 @@ const TeamMember = ({ user }: TeamMemberProps) => {
 
   return (
     <div
-      className={cn(
-        "relative box-content flex min-h-[350px] w-[300px] cursor-default flex-col items-center gap-3 rounded-2xl pt-12 text-center text-white max-lg:min-h-[380px] max-lg:w-[280px] max-sm:w-full max-sm:max-w-[300px] max-sm:min-h-[360px]",
-        hasOverlay && "group hover:-translate-y-1",
-      )}
+      className="relative box-content mx-auto flex min-h-[350px] w-[300px] cursor-default flex-col items-center gap-3 rounded-2xl pt-12 text-center text-white max-lg:min-h-[380px] max-lg:w-[280px] max-sm:min-h-[360px] max-sm:w-[min(100%,280px)]"
       key={user.profilePictureUrl}
       onMouseEnter={() => hasOverlay && setHovered(true)}
       onMouseLeave={() => setHovered(false)}
     >
       <div
         className={cn(
-          "relative z-[1] h-[225px] w-[225px] overflow-hidden rounded-full border-2 border-white/20 transition-[border-color,transform] duration-300 max-lg:h-[220px] max-lg:w-[220px] max-sm:h-[200px] max-sm:w-[200px]",
-          hasOverlay && "group-hover:scale-105 group-hover:border-primary-red",
+          "relative z-[1] h-[225px] w-[225px] overflow-hidden rounded-full border-2 border-white/20 transition-[border-color] duration-300 max-lg:h-[220px] max-lg:w-[220px] max-sm:h-[200px] max-sm:w-[200px]",
+          hovered && hasOverlay && "border-primary-red",
         )}
       >
         <Image
           alt="Headshot"
-          className="object-cover"
+          className="object-cover object-center"
           fill
           loading="eager"
           sizes={imageSize("teamHeadshot")}
@@ -71,7 +70,12 @@ const TeamMember = ({ user }: TeamMemberProps) => {
       </div>
 
       {hasOverlay && (
-        <div className="pointer-events-none absolute inset-0 z-10 flex items-center justify-center rounded-2xl bg-black/80 p-8 opacity-0 backdrop-blur-sm transition-opacity duration-300 group-hover:pointer-events-auto group-hover:opacity-100 max-lg:p-7 max-sm:rounded-xl max-sm:p-6">
+        <div
+          className={cn(
+            "pointer-events-none absolute inset-0 z-10 flex items-center justify-center rounded-2xl bg-black/80 p-8 backdrop-blur-sm transition-opacity duration-300 max-lg:p-7 max-sm:rounded-xl max-sm:p-6",
+            hovered ? "opacity-100" : "opacity-0",
+          )}
+        >
           <div className="flex w-full flex-col gap-6 overflow-hidden text-center">
             {overlayLeft && (
               <div className="text-base font-medium wrap-break-word text-white/95 italic after:mx-auto after:mt-2 after:block after:h-0.5 after:w-[150px] after:bg-[linear-gradient(90deg,transparent,var(--primary-red),transparent)] max-lg:text-[0.95rem] max-sm:text-sm">
@@ -85,7 +89,7 @@ const TeamMember = ({ user }: TeamMemberProps) => {
             )}
 
             {user.linkedIn && (
-              <div className="mx-auto flex h-8 w-8 items-center justify-center overflow-visible rounded bg-primary-red text-white transition-[transform,filter] duration-300 hover:scale-110 hover:drop-shadow-[0_4px_12px_rgba(245,23,32,0.6)] focus-visible:scale-110 focus-visible:drop-shadow-[0_4px_12px_rgba(245,23,32,0.6)]">
+              <div className="pointer-events-auto mx-auto flex h-8 w-8 items-center justify-center overflow-visible rounded bg-primary-red text-white transition-[transform,filter] duration-300 hover:scale-110 hover:drop-shadow-[0_4px_12px_rgba(245,23,32,0.6)] focus-visible:scale-110 focus-visible:drop-shadow-[0_4px_12px_rgba(245,23,32,0.6)]">
                 <Link
                   href={user.linkedIn}
                   rel="noopener noreferrer"
