--- src/lib/cars/fleet-timeline.ts
@@ -1,39 +1,0 @@
-import type { FleetTimelineEntry } from "./types";
-
-export const fleetTimeline: FleetTimelineEntry[] = [
-  {
-    name: "Schulich Helios",
-    years: "2026–present",
-    scrollTargetId: "Helios",
-  },
-  {
-    name: "Schulich Elysia",
-    years: "2019–2025",
-    scrollTargetId: "Elysia",
-  },
-  {
-    name: "Schulich Delta",
-    years: "2013–2015",
-    scrollTargetId: "Delta",
-  },
-  {
-    name: "Schulich Axiom",
-    years: "2010–2013",
-    scrollTargetId: "Axiom2011",
-  },
-  {
-    name: "Schulich I",
-    years: "2007–2010",
-    scrollTargetId: "Schulich1",
-  },
-  {
-    name: "Schulich Soleon",
-    years: "2004–2006",
-    scrollTargetId: "Soleon",
-  },
-  {
-    name: "X1 (prototype)",
-    years: "2005–2006",
-    scrollTargetId: "X1",
-  },
-];
