--- src/lib/cars/images.ts
@@ -1,17 +1,0 @@
-import backsplash1 from "public/assets/cars/backsplash1.png";
-import backsplash2 from "public/assets/cars/backsplash2.jpeg";
-import backsplash3 from "public/assets/cars/backsplash3.jpeg";
-import backsplash4 from "public/assets/cars/backsplash4.jpeg";
-import backsplash5 from "public/assets/cars/backsplash5.jpeg";
-import backsplash6 from "public/assets/cars/backsplash6.jpeg";
-import x1Image from "public/assets/cars/x1.png";
-
-export const carImages = {
-  helios: backsplash1,
-  elysia: backsplash2,
-  delta: backsplash3,
-  axiom: backsplash4,
-  schulich1: backsplash5,
-  soleon: backsplash6,
-  x1: x1Image,
-} as const;
